# api/storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "HS-SPORT API",
        "version": "1.0"
    },
    "basePath": "/api",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/gallery/calendar": {
            "post": {
                "tags": [
                    "Gallery"
                ],
                "summary": "Gallery Sports Calendar",
                "description": "Gallery Sports Calendar",
                "operationId": "api.gallery.calendar",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "seasonId",
                        "in": "query",
                        "description": "season ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "sportId",
                        "in": "query",
                        "description": "sport ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page number",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Show error message"
                    },
                    "401": {
                        "description": "Redirect to login page"
                    },
                    "409": {
                        "description": "Redirect to page 404"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
        "/game/upcoming": {
            "post": {
                "tags": [
                    "Game"
                ],
                "summary": "Upcoming game",
                "description": "Upcoming game",
                "operationId": "api.game.upcoming",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "sportId",
                        "in": "query",
                        "description": "1 or 2",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "genderId",
                        "in": "query",
                        "description": "1 or 2 or null",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "teamId",
                        "in": "query",
                        "description": "ID team",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "schoolId",
                        "in": "query",
                        "description": "ID school",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "schoolId",
                        "in": "query",
                        "description": "ID school",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "stateId",
                        "in": "query",
                        "description": "ID state",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "cityId",
                        "in": "query",
                        "description": "ID city",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Show error message"
                    },
                    "401": {
                        "description": "Redirect to login page"
                    },
                    "409": {
                        "description": "Redirect to page 404"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
        "/my-account": {
            "post": {
                "tags": [
                    "My Account"
                ],
                "summary": "My Account - index",
                "description": "My Account - index",
                "operationId": "api.my-account",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "token auth",
                        "required": true,
                        "type": "string",
                        "default": "Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpZCI6NzksImlzcyI6Imh0dHA6XC9cL2hzcy1hcGkudjEiLCJhdWQiOiJodHRwOlwvXC9oc3MtYXBpLnYxIiwiaWF0IjoxNTEzODg5MDAyLCJleHAiOjE1NDU0MjUwMDJ9.eB-Gcx4Dkko-C7Nfql0525nfOCJlGkLsl7m-GGhCPF4"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Show error message"
                    },
                    "401": {
                        "description": "Redirect to login page"
                    },
                    "409": {
                        "description": "Redirect to page 404"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/my-account/get-settings": {
            "post": {
                "tags": [
                    "My Account"
                ],
                "summary": "My Account - get settings",
                "description": "My Account - get settings",
                "operationId": "api.my-account.get-settings",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "token auth",
                        "required": true,
                        "type": "string",
                        "default": "Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpZCI6NzksImlzcyI6Imh0dHA6XC9cL2hzcy1hcGkudjEiLCJhdWQiOiJodHRwOlwvXC9oc3MtYXBpLnYxIiwiaWF0IjoxNTEzODg5MDAyLCJleHAiOjE1NDU0MjUwMDJ9.eB-Gcx4Dkko-C7Nfql0525nfOCJlGkLsl7m-GGhCPF4"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Show error message"
                    },
                    "401": {
                        "description": "Redirect to login page"
                    },
                    "409": {
                        "description": "Redirect to page 404"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/my-account/get-favorites": {
            "post": {
                "tags": [
                    "My Account"
                ],
                "summary": "My Account - get favorites",
                "description": "My Account - get favorites",
                "operationId": "api.my-account.get-favorites",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "token auth",
                        "required": true,
                        "type": "string",
                        "default": "Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpZCI6NzksImlzcyI6Imh0dHA6XC9cL2hzcy1hcGkudjEiLCJhdWQiOiJodHRwOlwvXC9oc3MtYXBpLnYxIiwiaWF0IjoxNTEzODg5MDAyLCJleHAiOjE1NDU0MjUwMDJ9.eB-Gcx4Dkko-C7Nfql0525nfOCJlGkLsl7m-GGhCPF4"
                    },
                    {
                        "name": "seasonId",
                        "in": "query",
                        "description": "season ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "modelType",
                        "in": "query",
                        "description": "player, team, school",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "sportId",
                        "in": "query",
                        "description": "sport ID",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page number",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Show error message"
                    },
                    "401": {
                        "description": "Redirect to login page"
                    },
                    "409": {
                        "description": "Redirect to page 404"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/my-account/get-events-calendar": {
            "post": {
                "tags": [
                    "My Account"
                ],
                "summary": "My Account - get events calendar",
                "description": "My Account - get events calendar",
                "operationId": "api.my-account.get-events-calendar",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "token auth",
                        "required": true,
                        "type": "string",
                        "default": "Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpZCI6NzksImlzcyI6Imh0dHA6XC9cL2hzcy1hcGkudjEiLCJhdWQiOiJodHRwOlwvXC9oc3MtYXBpLnYxIiwiaWF0IjoxNTEzODg5MDAyLCJleHAiOjE1NDU0MjUwMDJ9.eB-Gcx4Dkko-C7Nfql0525nfOCJlGkLsl7m-GGhCPF4"
                    },
                    {
                        "name": "calendarType",
                        "in": "query",
                        "description": "week || month || year",
                        "required": true,
                        "type": "string",
                        "default": "year"
                    },
                    {
                        "name": "startDate",
                        "in": "query",
                        "description": "Y-m-d",
                        "required": true,
                        "type": "string",
                        "default": "2016-01-01"
                    },
                    {
                        "name": "endDate",
                        "in": "query",
                        "description": "Y-m-d",
                        "required": true,
                        "type": "string",
                        "default": "2017-12-31"
                    },
                    {
                        "name": "sportId",
                        "in": "query",
                        "description": "sport ID",
                        "required": false,
                        "type": "integer",
                        "default": 1
                    },
                    {
                        "name": "modelType",
                        "in": "query",
                        "description": "game",
                        "required": true,
                        "type": "string",
                        "default": "game"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Show error message"
                    },
                    "401": {
                        "description": "Redirect to login page"
                    },
                    "409": {
                        "description": "Redirect to page 404"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
        "/scrapbook/get-scrapbook": {
            "post": {
                "tags": [
                    "Scrap Book Auth User"
                ],
                "summary": "Scrap Book - list",
                "description": "Scrap Book - list",
                "operationId": "api.scrapbook.get-scrapbook",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Show error message"
                    },
                    "401": {
                        "description": "Redirect to login page"
                    },
                    "409": {
                        "description": "Redirect to page 404"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
...
}